refactor(nextauth-options-refactor-to-utils/authoptions): nextauth options refactor to utils/authoptions
