+++ components/util/authoptions.ts
@@ -1,0 +1,99 @@
+import { type NextAuthOptions } from "next-auth";
+import CredentialsProvider from "next-auth/providers/credentials";
+
+const A_DAY = 86400; // a day
+const AuthOptions: NextAuthOptions = {
+  session: {
+    strategy: "jwt",
+    maxAge: A_DAY,
+  },
+
+  providers: [
+    CredentialsProvider({
+      // The name to display on the sign in form (e.g. 'Sign in with...')
+      name: "Credentials",
+      // The credentials is used to generate a suitable form on the sign in page.
+      // You can specify whatever fields you are expecting to be submitted.
+      // e.g. domain, username, password, 2FA token, etc.
+      // You can pass any HTML attribute to the <input> tag through the object.
+
+      // `credentials` is used to generate a form on the sign in page.
+      // You can specify which fields should be submitted, by adding keys to the `credentials` object.
+      // e.g. domain, username, password, 2FA token, etc.
+      // You can pass any HTML attribute to the <input> tag through the object.
+      credentials: {
+        email: { label: "email", type: "text", placeholder: "jsmith" },
+        password: { label: "Password", type: "password" },
+      },
+      // @ts-ignore
+      async authorize(credentials, req) {
+        // Add logic here to look up the user from the credentials supplied
+        console.log(credentials);
+
+        // @ts-ignore
+        const { email, password } = credentials;
+
+        const user: UserResponse = await fetch(
+          "https://store-hub-djxu.onrender.com/api/v1/auth/login",
+          {
+            method: "POST",
+            body: JSON.stringify({
+              email: credentials?.email,
+              password: credentials?.password,
+            }),
+            headers: { "Content-Type": "application/json" },
+          }
+        ).then((res) => res.json());
+        console.log(user);
+
+        if (!user?.error) {
+          const User = user.data.result.user;
+          // Any object returned will be saved in `user` property of the JWT
+          return {
+            user: User,
+            access_token: user.data.result.access_token,
+          };
+        } else {
+          // If you return null then an error will be displayed advising the user to check their details.
+          return null;
+
+          // You can also Reject this callback with an Error thus the user will be sent to the error page with the error message as a query parameter
+        }
+      },
+    }),
+  ],
+  pages: {
+    signIn: "/auth/signin",
+    signOut: "/auth/signout",
+  },
+  callbacks: {
+    session: ({ session, token }) => {
+      console.log("Session Callback", { session, token });
+      return {
+        ...session,
+        user: {
+          ...session.user,
+          user: token.user,
+          token: token.token,
+        },
+      };
+    },
+    jwt: ({ token, user }) => {
+      console.log("JWT Callback", { token, user });
+      if (user) {
+        const u = user as unknown as any;
+        return {
+          ...token,
+          user: u.user,
+          token: u.access_token,
+        };
+      }
+      return token;
+    },
+  },
+};
+
+const name = "uma victor";
+const age = 22;
+
+export { AuthOptions as default, name, age };
